Add pdfkit as a dependency
Quick search led me to:

https://stackoverflow.com/questions/742271/generating-pdf-files-with-javascript

And the main candidates are:

- jsPDF: https://github.com/MrRio/jsPDF
- PDFKit: https://github.com/foliojs/pdfkit
- pdfmake: http://pdfmake.org/

My 10-second evaluation of these suggests that PDFKit is the most
full-featured, and pdfmake is built on top of it; I may as well start
with PDFKit and if that doesn't work out I can look at one of the
alternatives.

So, `yarn add --dev pdfkit`, and 25 megabytes later, a `./node_modules`
folder is born.