Make the "blinking light demo" work
Pulling in js-yaml to read our YAML file, which is basically a
copy-pasta of the (out of date) content I currently have on my LinkedIn
profile, and feeding that into PDFKit to show that we can create a
document.

Run `npm run build` to spit out a "cv.pdf" file.