Create build directories
Separate public and private directories, because we're going to want to
make it straightforward and not error-prone to publish the former to the
gh-pages branch.

To avoid races, we do pre-emptive `mkdirSync` calls rather than
following a check-then-create pattern. If the `catch` swallows some more
sinister error here (like an permissions error) we'll just fail a bit
further on anyway.