Add gratuitous proxying to make language resolution transparent
I started writing this thinking of using proxies to do this dynamically
at property access time, and by the end of writing it realized that I
may as well just traverse the object once and substitute the language
entries eagerly. Committing this to show the intermediate state (which
works, but is needlessly complex).