Make sure PDFs don't change unless input changes
Works around: https://github.com/foliojs/pdfkit/issues/868

As described in that issue, if you generate the same PDF twice in a row
you'll get two different docs, and in my use case here (writing PDFs to
the "gh-pages" branch) that can produce noisy diffs.

The cause of the randomness can be found here:

https://github.com/foliojs/pdfkit/blob/2567227a92a10c0d88407dde232c4a3c3bc23d1a/lib/font/embedded.coffee#L103-L104

Work around it by (gasp) replacing `Math.random` with something
deterministic.

Now, we'll only get a different output when I change something that
affects the output. This means that to bump the "release version" of the
PDF (ie. the `ModDate`), I have to update the `version` property in the
YAML. I'm ok with that.