publish: properly show empty commits
When you create an empty commit on the gh-pages branch (for example, by
running `npm run publish` twice in a row, or by forgetting to run `npm
run build`), `git whatchanged` misbehaves (maybe, depending on your
point of view) compared to `git log`: the former skips the empty commit
and shows the parent commit instead, but `git log` will dutifully show
the empty commit.

Swap out `git whatchanged` for `git log --raw`, which has basically the
same output but doesn't skip empty commits. It's important that we show
empty commits because to give the user a chance to see and correct their
mistake (if it is a mistake; I can imagine you might want to push an
empty gh-pages commit to force a rebuild on the GitHub side). Ideally,
we would explode loudly for empty commits (the user could always just do
`git commit --allow-empty` if they really wanted), but I don't think
there is a built-in way to do that with `git-commit-tree`; to add such a
check I think I'd need to construct an explicit `git diff --quiet` test,
but I am not sure I can be bothered with that for now.

(Discovered this issue while testing changes to the `publish` script.)