Add "full" output variants
This one includes, for now, a new "Awards" section. I'm not going to
link to this from anywhere because I don't think the section is very
useful, and it also pushes the PDF version onto two pages (in the
"private" variants, which have a taller header, and also in the Spanhish
"public" variant), which I'll need to address.