Remove manual "collapse" hack
In writing the code to rollback unwanted pagebreaks (which involves
keeping a journal of all "commands" issued to build up the document), I
noticed I could get rid of the ugly manual "collapse" hack.

Whenever we emit a subheading, we can now automatically slurp up
preceding whitespace if the previous item was a heading.

The `args` property in the command journal isn't used in this
commit, but it will be in the future, when we use it to replay commands
after "rewinding" the output and inserting a page break in a better
place.