fix: make JSON schema validation work again
OK, so I lied a couple of commits ago (97ecf43) when I said I didn't
think I would be affected by breaking changes in the ajv package.

According to the docs here:

    https://ajv.js.org/guide/formats.html

you now need to install a separate package.

> From version 7 Ajv does not include formats defined by JSON Schema
> specification - these and several other formats are provided by
> ajv-formats plugin.

Another glorious triumph of modularization! My life is better for this
change.

With this change, all my `"format": "date"` and `"format": "email"`
(etc) schema elements work again, and we don't see this error during
build:

    Error: unknown format "date" ignored in schema at path "#/properties/version"

Still spewing out a bunch of warnings because "strict mode" is on by
default now, but the build still works, and I can deal with the errors
in a separate commit.